requirements.txt
fastapi[all]==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
jinja2==3.1.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
chromadb>=0.5.0
tiktoken>=0.5.2
pypdf==3.17.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
chardet>=5.2.0
